Change GrandTotal type from double to decimal
Updated the GrandTotal property in AdminOrderIndexViewModel from double to decimal to enhance precision and accuracy in financial calculations.

+++ ST10361554-CLDV6212-POE-Part-3-WebApp/ViewModels/AdminOrderIndexViewModel.cs
@@ -12,7 +12,7 @@
         public required string OrderStatus { get; set; }
 
         // Total amount for the order
-        public double GrandTotal { get; set; }
+        public decimal GrandTotal { get; set; }
 
         // Unique identifier for the customer who placed the order
         public required string CustomerID { get; set; }
